requirements.txt
pyTelegramBotAPI
python-dotenv
watchdog
opencv-python